+++ KlayGE/Plugins/Src/Render/D3D11/D3D11TextureCube.cpp
@@ -94,7 +94,7 @@ namespace KlayGE
 
 	uint32_t D3D11TextureCube::Width(int level) const
 	{
-		BOOST_ASSERT(level < numMipMaps_);
+		BOOST_ASSERT(static_cast<uint32_t>(level) < numMipMaps_);
 
 		return widthes_[level];
 	}
@@ -107,7 +107,7 @@ namespace KlayGE
 	uint32_t D3D11TextureCube::Depth(int level) const
 	{
 		UNREF_PARAM(level);
-		BOOST_ASSERT(level < numMipMaps_);
+		BOOST_ASSERT(static_cast<uint32_t>(level) < numMipMaps_);
 
 		return 1;
 	}
