+++ KlayGE/Plugins/Src/Render/D3D9/D3D9Texture2D.cpp
@@ -130,14 +130,14 @@ namespace KlayGE
 
 	uint32_t D3D9Texture2D::Width(int level) const
 	{
-		BOOST_ASSERT(level < numMipMaps_);
+		BOOST_ASSERT(static_cast<uint32_t>(level) < numMipMaps_);
 
 		return widths_[level];
 	}
 
 	uint32_t D3D9Texture2D::Height(int level) const
 	{
-		BOOST_ASSERT(level < numMipMaps_);
+		BOOST_ASSERT(static_cast<uint32_t>(level) < numMipMaps_);
 
 		return heights_[level];
 	}
@@ -145,7 +145,7 @@ namespace KlayGE
 	uint32_t D3D9Texture2D::Depth(int level) const
 	{
 		UNREF_PARAM(level);
-		BOOST_ASSERT(level < numMipMaps_);
+		BOOST_ASSERT(static_cast<uint32_t>(level) < numMipMaps_);
 
 		return 1;
 	}
