+++ KlayGE/Plugins/Src/Render/OpenGL/OGLTexture1D.cpp
@@ -146,7 +146,7 @@ namespace KlayGE
 
 	uint32_t OGLTexture1D::Width(int level) const
 	{
-		BOOST_ASSERT(level < numMipMaps_);
+		BOOST_ASSERT(static_cast<uint32_t>(level) < numMipMaps_);
 
 		return widthes_[level];
 	}
@@ -154,7 +154,7 @@ namespace KlayGE
 	uint32_t OGLTexture1D::Height(int level) const
 	{
 		UNREF_PARAM(level);
-		BOOST_ASSERT(level < numMipMaps_);
+		BOOST_ASSERT(static_cast<uint32_t>(level) < numMipMaps_);
 
 		return 1;
 	}
@@ -162,7 +162,7 @@ namespace KlayGE
 	uint32_t OGLTexture1D::Depth(int level) const
 	{
 		UNREF_PARAM(level);
-		BOOST_ASSERT(level < numMipMaps_);
+		BOOST_ASSERT(static_cast<uint32_t>(level) < numMipMaps_);
 
 		return 1;
 	}
