+++ KlayGE/Plugins/Src/Render/OpenGL/OGLTexture2D.cpp
@@ -152,14 +152,14 @@ namespace KlayGE
 
 	uint32_t OGLTexture2D::Width(int level) const
 	{
-		BOOST_ASSERT(level < numMipMaps_);
+		BOOST_ASSERT(static_cast<uint32_t>(level) < numMipMaps_);
 
 		return widthes_[level];
 	}
 
 	uint32_t OGLTexture2D::Height(int level) const
 	{
-		BOOST_ASSERT(level < numMipMaps_);
+		BOOST_ASSERT(static_cast<uint32_t>(level) < numMipMaps_);
 
 		return heights_[level];
 	}
@@ -167,7 +167,7 @@ namespace KlayGE
 	uint32_t OGLTexture2D::Depth(int level) const
 	{
 		UNREF_PARAM(level);
-		BOOST_ASSERT(level < numMipMaps_);
+		BOOST_ASSERT(static_cast<uint32_t>(level) < numMipMaps_);
 
 		return 1;
 	}
